+++ Assets/01.Scripts/OutGame/Goods/Repository/CurrencySaveData.cs
@@ -3,7 +3,7 @@
     public class CurrencySaveData
     {
         // 재화 
-        public int Currency;
+        public long Currency;
